removed unused variables to stop warnings

+++ scpi.hpp
@@ -81,7 +81,7 @@ public:
         std::string full_message = message + delimiter;
 
 
-        bool sentBytes = NET_WriteToStreamSocket(clientTCPSocket, (void*)full_message.c_str(), strlen(full_message.c_str()));
+        bool sentBytes = NET_WriteToStreamSocket(clientTCPSocket, (void*)full_message.c_str(),static_cast<int>( strlen(full_message.c_str())));
 
         //ADD IN CASE DATA ARE NOT SENT PROPERLY : NET_WaitUntilStreamSocketDrained
